+++ src/user/user.service.ts
@@ -6,7 +6,7 @@ import {
 import { Repository } from 'typeorm';
 import { InjectRepository } from '@nestjs/typeorm';
 import { UserEntity } from 'src/entity/user.entity';
-import CreateUserRequestDTO from './dto/create-user-request.dto';
+import CreateUserRequestDTO from './dto/user-request.dto';
 import { UserResponseDTO } from './dto/user-response.dto';
 import { SSORequestDTO } from './dto/sso-request.dto';
